requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pillow==11.0.0
pydantic==2.10.0
httpx==0.27.0
python-dotenv==1.0.1
sse-starlette==2.1.0
opencv-python-headless==4.9.0.80
numpy==1.26.4

# Testing dependencies
pytest==9.0.2
pytest-asyncio==1.3.0
pytest-cov==7.0.0
hypothesis==6.148.8
